+++ modules/upload_sample.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+# STARTFILES
+# targets.txt "Targets_description"
+# ENDFILES
+# ARGS
+# NSE "NSE_Query_String"
+# ENDARGS
+# AUTHOR:
+cat /tmp/targets.txt
+echo
+echo $1